+++ src/bll/setting.go
@@ -437,6 +437,64 @@ func (b *Setting) ListUsers(ctx context.Context, productName, moduleName, settin
 	return res, nil
 }
 
+// RollbackUserSetting ...
+func (b *Setting) RollbackUserSetting(ctx context.Context, productName, moduleName, settingName, uid string) (*tpl.BoolRes, error) {
+	productID, err := b.ms.Product.AcquireID(ctx, productName)
+	if err != nil {
+		return nil, err
+	}
+
+	module, err := b.ms.Module.Acquire(ctx, productID, moduleName)
+	if err != nil {
+		return nil, err
+	}
+
+	setting, err := b.ms.Setting.Acquire(ctx, module.ID, settingName)
+	if err != nil {
+		return nil, err
+	}
+
+	user, err := b.ms.User.Acquire(ctx, uid)
+	if err != nil {
+		return nil, err
+	}
+
+	err = b.ms.Setting.RollbackUserSetting(ctx, user.ID, setting.ID)
+	if err != nil {
+		return nil, err
+	}
+	return &tpl.BoolRes{Result: true}, nil
+}
+
+// DeleteUser ...
+func (b *Setting) DeleteUser(ctx context.Context, productName, moduleName, settingName, uid string) (*tpl.BoolRes, error) {
+	productID, err := b.ms.Product.AcquireID(ctx, productName)
+	if err != nil {
+		return nil, err
+	}
+
+	module, err := b.ms.Module.Acquire(ctx, productID, moduleName)
+	if err != nil {
+		return nil, err
+	}
+
+	setting, err := b.ms.Setting.Acquire(ctx, module.ID, settingName)
+	if err != nil {
+		return nil, err
+	}
+
+	user, err := b.ms.User.Acquire(ctx, uid)
+	if err != nil {
+		return nil, err
+	}
+
+	rowsAffected, err := b.ms.Setting.RemoveUserSetting(ctx, user.ID, setting.ID)
+	if err != nil {
+		return nil, err
+	}
+	return &tpl.BoolRes{Result: rowsAffected > 0}, nil
+}
+
 // ListGroups 返回产品下功能配置项的群组列表
 func (b *Setting) ListGroups(ctx context.Context, productName, moduleName, settingName string, pg tpl.Pagination) (*tpl.SettingGroupsInfoRes, error) {
 	productID, err := b.ms.Product.AcquireID(ctx, productName)
@@ -465,4 +523,62 @@ func (b *Setting) ListGroups(ctx context.Context, productName, moduleName, setti
 		res.Result = res.Result[:pg.PageSize]
 	}
 	return res, nil
+}
+
+// RollbackGroupSetting ...
+func (b *Setting) RollbackGroupSetting(ctx context.Context, productName, moduleName, settingName, uid string) (*tpl.BoolRes, error) {
+	productID, err := b.ms.Product.AcquireID(ctx, productName)
+	if err != nil {
+		return nil, err
+	}
+
+	module, err := b.ms.Module.Acquire(ctx, productID, moduleName)
+	if err != nil {
+		return nil, err
+	}
+
+	setting, err := b.ms.Setting.Acquire(ctx, module.ID, settingName)
+	if err != nil {
+		return nil, err
+	}
+
+	group, err := b.ms.Group.Acquire(ctx, uid)
+	if err != nil {
+		return nil, err
+	}
+
+	err = b.ms.Setting.RollbackGroupSetting(ctx, group.ID, setting.ID)
+	if err != nil {
+		return nil, err
+	}
+	return &tpl.BoolRes{Result: true}, nil
+}
+
+// DeleteGroup ...
+func (b *Setting) DeleteGroup(ctx context.Context, productName, moduleName, settingName, uid string) (*tpl.BoolRes, error) {
+	productID, err := b.ms.Product.AcquireID(ctx, productName)
+	if err != nil {
+		return nil, err
+	}
+
+	module, err := b.ms.Module.Acquire(ctx, productID, moduleName)
+	if err != nil {
+		return nil, err
+	}
+
+	setting, err := b.ms.Setting.Acquire(ctx, module.ID, settingName)
+	if err != nil {
+		return nil, err
+	}
+
+	group, err := b.ms.Group.Acquire(ctx, uid)
+	if err != nil {
+		return nil, err
+	}
+
+	rowsAffected, err := b.ms.Setting.RemoveGroupSetting(ctx, group.ID, setting.ID)
+	if err != nil {
+		return nil, err
+	}
+	return &tpl.BoolRes{Result: rowsAffected > 0}, nil
 }